karmada-metrics-adapter: reduce memory usage
When there is a large amount of pod usage in the member cluster, metrics-adapter will consume a lot of memory. The reason is that it caches all the information of all pods in the cluster. However, we don't need all this information, so we trim some of the information to reduce memory usage.

Signed-off-by: chaunceyjiang <chaunceyjiang@gmail.com>